+++ src/data_source/fetching.rs
@@ -80,16 +80,12 @@ use crate::{
         PayloadQueryData, QueryablePayload, TransactionHash, TransactionIndex,
         UpdateAvailabilityData, VidCommonQueryData,
     },
-    fetching::{
-        self,
-        request::{self, PayloadRequest},
-        Callback, Provider,
-    },
+    fetching::{self, request, Provider},
     metrics::PrometheusMetrics,
     node::{NodeDataSource, SyncStatus},
     status::StatusDataSource,
     task::BackgroundTask,
-    Header, Payload, QueryResult, SignatureKey, VidCommon, VidShare,
+    Payload, QueryResult, SignatureKey, VidShare,
 };
 use anyhow::Context;
 use async_std::{
@@ -98,21 +94,30 @@ use async_std::{
 };
 use async_trait::async_trait;
 use derivative::Derivative;
-use derive_more::{Display, From};
+use derive_more::From;
 use futures::{
     future::{self, join_all, BoxFuture, FutureExt},
     stream::{self, BoxStream, Stream, StreamExt},
 };
-use hotshot_types::traits::{block_contents::BlockHeader, node_implementation::NodeType};
+use hotshot_types::traits::node_implementation::NodeType;
 use std::{
-    cmp::{min, Ordering},
+    cmp::min,
     fmt::{Debug, Display},
-    future::IntoFuture,
-    iter::once,
     marker::PhantomData,
     ops::{Bound, Deref, DerefMut, Range, RangeBounds},
     time::Duration,
 };
+
+mod block;
+mod header;
+mod leaf;
+mod transaction;
+mod vid;
+
+use block::PayloadFetcher;
+use leaf::LeafFetcher;
+use transaction::TransactionRequest;
+use vid::{VidCommonFetcher, VidCommonRequest};
 
 /// Builder for [`FetchingDataSource`] with configuration.
 pub struct Builder<Types, S, P> {
@@ -424,7 +429,7 @@ where
     where
         ID: Into<BlockId<Types>> + Send + Sync,
     {
-        self.fetcher.get(VidCommonRequest(id.into())).await
+        self.fetcher.get(VidCommonRequest::from(id.into())).await
     }
 
     async fn get_leaf_range<R>(&self, range: R) -> Self::LeafRange<R>
@@ -459,7 +464,7 @@ where
         &self,
         hash: TransactionHash<Types>,
     ) -> Fetch<(BlockQueryData<Types>, TransactionIndex<Types>)> {
-        self.fetcher.get(TransactionRequest(hash)).await
+        self.fetcher.get(TransactionRequest::from(hash)).await
     }
 }
 
@@ -564,10 +569,9 @@ where
 {
     storage: RwLock<NotifyStorage<Types, S>>,
     provider: Arc<P>,
-    payload_fetcher: Arc<fetching::Fetcher<request::PayloadRequest, PayloadCallback<Types, S, P>>>,
-    leaf_fetcher: Arc<fetching::Fetcher<request::LeafRequest, LeafCallback<Types, S, P>>>,
-    vid_common_fetcher:
-        Arc<fetching::Fetcher<request::VidCommonRequest, VidCommonCallback<Types, S, P>>>,
+    payload_fetcher: Arc<PayloadFetcher<Types, S, P>>,
+    leaf_fetcher: Arc<LeafFetcher<Types, S, P>>,
+    vid_common_fetcher: Arc<VidCommonFetcher<Types, S, P>>,
     range_chunk_size: usize,
 }
 
@@ -1081,914 +1085,6 @@ where
         R: RangeBounds<usize> + Send + 'static;
 
     fn height(&self) -> usize;
-}
-
-impl<Types> FetchRequest for LeafId<Types>
-where
-    Types: NodeType,
-{
-    fn might_exist(self, block_height: usize) -> bool {
-        if let LeafId::Number(n) = self {
-            n < block_height
-        } else {
-            true
-        }
-    }
-}
-
-#[async_trait]
-impl<Types> Fetchable<Types> for LeafQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type Request = LeafId<Types>;
-
-    fn satisfies(&self, req: Self::Request) -> bool {
-        match req {
-            LeafId::Number(n) => self.height() == n as u64,
-            LeafId::Hash(h) => self.hash() == h,
-        }
-    }
-
-    async fn passive_fetch<S>(
-        storage: &NotifyStorage<Types, S>,
-        req: Self::Request,
-    ) -> BoxFuture<'static, Option<Self>>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage
-            .leaf_notifier
-            .wait_for(move |leaf| leaf.satisfies(req))
-            .await
-            .into_future()
-            .boxed()
-    }
-
-    async fn active_fetch<S, P>(
-        fetcher: Arc<Fetcher<Types, S, P>>,
-        _storage: &RwLockReadGuard<'_, NotifyStorage<Types, S>>,
-        req: Self::Request,
-    ) where
-        S: AvailabilityStorage<Types> + 'static,
-        P: AvailabilityProvider<Types>,
-    {
-        fetch_leaf_with_callbacks(fetcher, req, None)
-    }
-
-    async fn load<S>(storage: &NotifyStorage<Types, S>, req: Self::Request) -> QueryResult<Self>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage.storage.get_leaf(req).await
-    }
-}
-
-fn fetch_leaf_with_callbacks<Types, S, P, I>(
-    fetcher: Arc<Fetcher<Types, S, P>>,
-    req: LeafId<Types>,
-    callbacks: I,
-) where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types> + 'static,
-    P: AvailabilityProvider<Types>,
-    I: IntoIterator<Item = LeafCallback<Types, S, P>> + Send + 'static,
-    I::IntoIter: Send,
-{
-    match req {
-        LeafId::Number(n) => {
-            let fetcher = fetcher.clone();
-            fetcher.leaf_fetcher.clone().spawn_fetch(
-                n.into(),
-                fetcher.provider.clone(),
-                once(LeafCallback::Leaf { fetcher }).chain(callbacks),
-            );
-        }
-        LeafId::Hash(h) => {
-            // We don't actively fetch leaves when requested by hash, because we have no way of
-            // knowing whether a leaf with such a hash actually exists, and we don't want to bother
-            // peers with requests for non-existant leaves.
-            tracing::debug!("not fetching unknown leaf {h}");
-        }
-    }
-}
-
-async fn store_leaf<Types, S>(
-    storage: &mut NotifyStorage<Types, S>,
-    leaf: LeafQueryData<Types>,
-) -> anyhow::Result<()>
-where
-    Types: NodeType,
-    S: UpdateAvailabilityData<Types> + VersionedDataSource,
-{
-    storage.insert_leaf(leaf).await?;
-    storage.commit().await?;
-    Ok(())
-}
-
-#[async_trait]
-impl<Types> RangedFetchable<Types> for LeafQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type RangedRequest = LeafId<Types>;
-
-    async fn load_range<S, R>(
-        storage: &NotifyStorage<Types, S>,
-        range: R,
-    ) -> QueryResult<Vec<QueryResult<Self>>>
-    where
-        S: AvailabilityStorage<Types>,
-        R: RangeBounds<usize> + Send + 'static,
-    {
-        storage.storage.get_leaf_range(range).await
-    }
-
-    fn height(&self) -> usize {
-        self.height() as usize
-    }
-}
-
-/// A request to fetch a block.
-///
-/// Blocks can be requested either directly by their [`BlockId`], or indirectly, by requesting a
-/// block containing a particular transaction.
-#[derive(Derivative, From, Display)]
-#[derivative(Ord = "feature_allow_slow_enum")]
-#[derivative(
-    Copy(bound = ""),
-    Debug(bound = ""),
-    PartialEq(bound = ""),
-    Eq(bound = ""),
-    Ord(bound = ""),
-    Hash(bound = "")
-)]
-pub enum BlockRequest<Types>
-where
-    Types: NodeType,
-{
-    Id(BlockId<Types>),
-    WithTransaction(TransactionHash<Types>),
-}
-
-impl<Types> From<usize> for BlockRequest<Types>
-where
-    Types: NodeType,
-{
-    fn from(i: usize) -> Self {
-        Self::Id(i.into())
-    }
-}
-
-impl<Types> Clone for BlockRequest<Types>
-where
-    Types: NodeType,
-{
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<Types> PartialOrd for BlockRequest<Types>
-where
-    Types: NodeType,
-{
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl<Types> FetchRequest for BlockId<Types>
-where
-    Types: NodeType,
-{
-    fn might_exist(self, block_height: usize) -> bool {
-        if let BlockId::Number(n) = self {
-            n < block_height
-        } else {
-            true
-        }
-    }
-}
-
-impl<Types> FetchRequest for BlockRequest<Types>
-where
-    Types: NodeType,
-{
-    fn might_exist(self, block_height: usize) -> bool {
-        if let BlockRequest::Id(id) = self {
-            id.might_exist(block_height)
-        } else {
-            true
-        }
-    }
-}
-
-#[async_trait]
-impl<Types> Fetchable<Types> for BlockQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type Request = BlockRequest<Types>;
-
-    fn satisfies(&self, req: Self::Request) -> bool {
-        match req {
-            BlockRequest::Id(BlockId::Number(n)) => self.height() == n as u64,
-            BlockRequest::Id(BlockId::Hash(h)) => self.hash() == h,
-            BlockRequest::Id(BlockId::PayloadHash(h)) => self.payload_hash() == h,
-            BlockRequest::WithTransaction(h) => self.transaction_by_hash(h).is_some(),
-        }
-    }
-
-    async fn passive_fetch<S>(
-        storage: &NotifyStorage<Types, S>,
-        req: Self::Request,
-    ) -> BoxFuture<'static, Option<Self>>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage
-            .block_notifier
-            .wait_for(move |block| block.satisfies(req))
-            .await
-            .into_future()
-            .boxed()
-    }
-
-    async fn active_fetch<S, P>(
-        fetcher: Arc<Fetcher<Types, S, P>>,
-        storage: &RwLockReadGuard<'_, NotifyStorage<Types, S>>,
-        req: Self::Request,
-    ) where
-        S: AvailabilityStorage<Types> + 'static,
-        P: AvailabilityProvider<Types>,
-    {
-        fetch_header_and_then(storage, req, HeaderCallback::Payload { fetcher }).await
-    }
-
-    async fn load<S>(storage: &NotifyStorage<Types, S>, req: Self::Request) -> QueryResult<Self>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        match req {
-            BlockRequest::Id(id) => storage.storage.get_block(id).await,
-            BlockRequest::WithTransaction(h) => {
-                Ok(storage.storage.get_block_with_transaction(h).await?.0)
-            }
-        }
-    }
-}
-
-#[async_trait]
-impl<Types> RangedFetchable<Types> for BlockQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type RangedRequest = BlockRequest<Types>;
-
-    async fn load_range<S, R>(
-        storage: &NotifyStorage<Types, S>,
-        range: R,
-    ) -> QueryResult<Vec<QueryResult<Self>>>
-    where
-        S: AvailabilityStorage<Types>,
-        R: RangeBounds<usize> + Send + 'static,
-    {
-        storage.storage.get_block_range(range).await
-    }
-
-    fn height(&self) -> usize {
-        self.height() as usize
-    }
-}
-
-fn fetch_block_with_header<Types, S, P>(fetcher: Arc<Fetcher<Types, S, P>>, header: Header<Types>)
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types> + 'static,
-    P: AvailabilityProvider<Types>,
-{
-    // Now that we have the header, we only need to retrieve the payload.
-    tracing::info!(
-        "spawned active fetch for payload {:?} (height {})",
-        header.payload_commitment(),
-        header.block_number()
-    );
-    fetcher.payload_fetcher.spawn_fetch(
-        PayloadRequest(header.payload_commitment()),
-        fetcher.provider.clone(),
-        once(PayloadCallback {
-            header,
-            fetcher: fetcher.clone(),
-        }),
-    );
-}
-
-async fn store_block<Types, S>(
-    storage: &mut NotifyStorage<Types, S>,
-    block: BlockQueryData<Types>,
-) -> anyhow::Result<()>
-where
-    Types: NodeType,
-    S: UpdateAvailabilityData<Types> + VersionedDataSource,
-{
-    storage.insert_block(block).await?;
-    storage.commit().await?;
-    Ok(())
-}
-
-#[derive(Clone, Copy, Debug)]
-struct TransactionRequest<Types: NodeType>(TransactionHash<Types>);
-
-impl<Types: NodeType> FetchRequest for TransactionRequest<Types> {}
-
-#[async_trait]
-impl<Types> Fetchable<Types> for (BlockQueryData<Types>, TransactionIndex<Types>)
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type Request = TransactionRequest<Types>;
-
-    fn satisfies(&self, req: Self::Request) -> bool {
-        self.0.transaction_by_hash(req.0).is_some()
-    }
-
-    async fn passive_fetch<S>(
-        storage: &NotifyStorage<Types, S>,
-        req: Self::Request,
-    ) -> BoxFuture<'static, Option<Self>>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        let wait_block = storage
-            .block_notifier
-            .wait_for(move |block| block.satisfies(req.0.into()))
-            .await;
-
-        async move {
-            let block = wait_block.await?;
-
-            // This `unwrap` is safe, `wait_for` only returns blocks which satisfy the request, and
-            // in this case that means the block must contain the requested transaction.
-            let ix = block.transaction_by_hash(req.0).unwrap();
-
-            Some((block, ix))
-        }
-        .boxed()
-    }
-
-    async fn active_fetch<S, P>(
-        _fetcher: Arc<Fetcher<Types, S, P>>,
-        _storage: &RwLockReadGuard<'_, NotifyStorage<Types, S>>,
-        req: Self::Request,
-    ) where
-        S: AvailabilityStorage<Types> + 'static,
-        P: AvailabilityProvider<Types>,
-    {
-        // We don't actively fetch blocks when requested by transaction, because without the block
-        // payload, we have no way of knowing whether a block with such a transaction actually
-        // exists, and we don't want to bother peers with requests for non-existant blocks.
-        tracing::debug!("not fetching block with unknown transaction {req:?}");
-    }
-
-    async fn load<S>(storage: &NotifyStorage<Types, S>, req: Self::Request) -> QueryResult<Self>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage.storage.get_block_with_transaction(req.0).await
-    }
-}
-
-#[async_trait]
-impl<Types> Fetchable<Types> for PayloadQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type Request = BlockId<Types>;
-
-    fn satisfies(&self, req: Self::Request) -> bool {
-        match req {
-            BlockId::Number(n) => self.height() == n as u64,
-            BlockId::Hash(h) => self.block_hash() == h,
-            BlockId::PayloadHash(h) => self.hash() == h,
-        }
-    }
-
-    async fn passive_fetch<S>(
-        storage: &NotifyStorage<Types, S>,
-        req: Self::Request,
-    ) -> BoxFuture<'static, Option<Self>>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage
-            .block_notifier
-            .wait_for(move |block| block.satisfies(req.into()))
-            .await
-            .into_future()
-            .map(|block| block.map(PayloadQueryData::from))
-            .boxed()
-    }
-
-    async fn active_fetch<S, P>(
-        fetcher: Arc<Fetcher<Types, S, P>>,
-        storage: &RwLockReadGuard<'_, NotifyStorage<Types, S>>,
-        req: Self::Request,
-    ) where
-        S: AvailabilityStorage<Types> + 'static,
-        P: AvailabilityProvider<Types>,
-    {
-        // We don't have storage for the payload alone, only the whole block. So if we need to fetch
-        // the payload, we just fetch the whole block (which may end up fetching only the payload,
-        // if that's all that's needed to complete the block).
-        BlockQueryData::active_fetch(fetcher, storage, req.into()).await
-    }
-
-    async fn load<S>(storage: &NotifyStorage<Types, S>, req: Self::Request) -> QueryResult<Self>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage.storage.get_payload(req).await
-    }
-}
-
-#[async_trait]
-impl<Types> RangedFetchable<Types> for PayloadQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type RangedRequest = BlockId<Types>;
-
-    async fn load_range<S, R>(
-        storage: &NotifyStorage<Types, S>,
-        range: R,
-    ) -> QueryResult<Vec<QueryResult<Self>>>
-    where
-        S: AvailabilityStorage<Types>,
-        R: RangeBounds<usize> + Send + 'static,
-    {
-        storage.storage.get_payload_range(range).await
-    }
-
-    fn height(&self) -> usize {
-        self.height() as usize
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-struct VidCommonRequest<Types: NodeType>(BlockId<Types>);
-
-impl<Types: NodeType> From<usize> for VidCommonRequest<Types> {
-    fn from(n: usize) -> Self {
-        Self(n.into())
-    }
-}
-
-impl<Types> FetchRequest for VidCommonRequest<Types>
-where
-    Types: NodeType,
-{
-    fn might_exist(self, block_height: usize) -> bool {
-        if let BlockId::Number(n) = self.0 {
-            n > 0 && n < block_height
-        } else {
-            true
-        }
-    }
-}
-
-#[async_trait]
-impl<Types> Fetchable<Types> for VidCommonQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type Request = VidCommonRequest<Types>;
-
-    fn satisfies(&self, req: Self::Request) -> bool {
-        match req.0 {
-            BlockId::Number(n) => self.height() == n as u64,
-            BlockId::Hash(h) => self.block_hash() == h,
-            BlockId::PayloadHash(h) => self.payload_hash() == h,
-        }
-    }
-
-    async fn passive_fetch<S>(
-        storage: &NotifyStorage<Types, S>,
-        req: Self::Request,
-    ) -> BoxFuture<'static, Option<Self>>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage
-            .vid_common_notifier
-            .wait_for(move |vid| vid.satisfies(req))
-            .await
-            .into_future()
-            .boxed()
-    }
-
-    async fn active_fetch<S, P>(
-        fetcher: Arc<Fetcher<Types, S, P>>,
-        storage: &RwLockReadGuard<'_, NotifyStorage<Types, S>>,
-        req: Self::Request,
-    ) where
-        S: AvailabilityStorage<Types> + 'static,
-        P: AvailabilityProvider<Types>,
-    {
-        fetch_header_and_then(storage, req.0.into(), HeaderCallback::VidCommon { fetcher }).await
-    }
-
-    async fn load<S>(storage: &NotifyStorage<Types, S>, req: Self::Request) -> QueryResult<Self>
-    where
-        S: AvailabilityStorage<Types>,
-    {
-        storage.storage.get_vid_common(req.0).await
-    }
-}
-
-#[async_trait]
-impl<Types> RangedFetchable<Types> for VidCommonQueryData<Types>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-{
-    type RangedRequest = VidCommonRequest<Types>;
-
-    async fn load_range<S, R>(
-        storage: &NotifyStorage<Types, S>,
-        range: R,
-    ) -> QueryResult<Vec<QueryResult<Self>>>
-    where
-        S: AvailabilityStorage<Types>,
-        R: RangeBounds<usize> + Send + 'static,
-    {
-        storage.storage.get_vid_common_range(range).await
-    }
-
-    fn height(&self) -> usize {
-        self.height() as usize
-    }
-}
-
-fn fetch_vid_common_with_header<Types, S, P>(
-    fetcher: Arc<Fetcher<Types, S, P>>,
-    header: Header<Types>,
-) where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types> + 'static,
-    P: AvailabilityProvider<Types>,
-{
-    // Now that we have the header, we only need to retrieve the VID common data.
-    tracing::info!(
-        "spawned active fetch for VID common {:?} (height {})",
-        header.payload_commitment(),
-        header.block_number()
-    );
-    fetcher.vid_common_fetcher.spawn_fetch(
-        request::VidCommonRequest(header.payload_commitment()),
-        fetcher.provider.clone(),
-        once(VidCommonCallback {
-            header,
-            fetcher: fetcher.clone(),
-        }),
-    );
-}
-
-async fn store_vid_common<Types, S>(
-    storage: &mut NotifyStorage<Types, S>,
-    common: VidCommonQueryData<Types>,
-) -> anyhow::Result<()>
-where
-    Types: NodeType,
-    S: UpdateAvailabilityData<Types> + VersionedDataSource,
-{
-    storage.insert_vid(common, None).await?;
-    storage.commit().await?;
-    Ok(())
-}
-
-#[derive(Derivative)]
-#[derivative(Debug(bound = ""))]
-struct PayloadCallback<Types: NodeType, S, P> {
-    header: Header<Types>,
-    #[derivative(Debug = "ignore")]
-    fetcher: Arc<Fetcher<Types, S, P>>,
-}
-
-impl<Types: NodeType, S, P> PartialEq for PayloadCallback<Types, S, P> {
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other).is_eq()
-    }
-}
-
-impl<Types: NodeType, S, P> Eq for PayloadCallback<Types, S, P> {}
-
-impl<Types: NodeType, S, P> Ord for PayloadCallback<Types, S, P> {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.header.block_number().cmp(&other.header.block_number())
-    }
-}
-
-impl<Types: NodeType, S, P> PartialOrd for PayloadCallback<Types, S, P> {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl<Types: NodeType, S, P> Callback<Payload<Types>> for PayloadCallback<Types, S, P>
-where
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types>,
-    P: AvailabilityProvider<Types>,
-{
-    async fn run(self, payload: Payload<Types>) {
-        tracing::info!("fetched payload {:?}", self.header.payload_commitment());
-        let block = BlockQueryData::new(self.header, payload);
-        let height = block.height();
-
-        // Store the block in local storage, so we can avoid fetching it in the future.
-        {
-            let mut storage = self.fetcher.storage.write().await;
-            if let Err(err) = store_block(&mut *storage, block).await {
-                // It is unfortunate if this fails, but we can still proceed by returning
-                // the block that we fetched, keeping it in memory. Simply log the error and
-                // move on.
-                tracing::warn!("failed to store fetched block {height}: {err}");
-            }
-        }
-    }
-}
-
-#[derive(Derivative)]
-#[derivative(Debug(bound = ""))]
-struct VidCommonCallback<Types: NodeType, S, P> {
-    header: Header<Types>,
-    #[derivative(Debug = "ignore")]
-    fetcher: Arc<Fetcher<Types, S, P>>,
-}
-
-impl<Types: NodeType, S, P> PartialEq for VidCommonCallback<Types, S, P> {
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other).is_eq()
-    }
-}
-
-impl<Types: NodeType, S, P> Eq for VidCommonCallback<Types, S, P> {}
-
-impl<Types: NodeType, S, P> Ord for VidCommonCallback<Types, S, P> {
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.header.block_number().cmp(&other.header.block_number())
-    }
-}
-
-impl<Types: NodeType, S, P> PartialOrd for VidCommonCallback<Types, S, P> {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl<Types: NodeType, S, P> Callback<VidCommon> for VidCommonCallback<Types, S, P>
-where
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types>,
-    P: AvailabilityProvider<Types>,
-{
-    async fn run(self, common: VidCommon) {
-        tracing::info!("fetched VID common {:?}", self.header.payload_commitment());
-        let common = VidCommonQueryData::new(self.header, common);
-        let height = common.height();
-
-        // Store the data in local storage, so we can avoid fetching it in the future.
-        {
-            let mut storage = self.fetcher.storage.write().await;
-            if let Err(err) = store_vid_common(&mut *storage, common).await {
-                // It is unfortunate if this fails, but we can still proceed by returning
-                // the block that we fetched, keeping it in memory. Simply log the error and
-                // move on.
-                tracing::warn!("failed to store fetched VID common {height}: {err}");
-            }
-        }
-    }
-}
-
-#[derive(Derivative, From)]
-#[derivative(Debug(bound = ""))]
-enum LeafCallback<Types: NodeType, S, P> {
-    /// Callback when fetching the leaf for its own sake.
-    #[from(ignore)]
-    Leaf {
-        #[derivative(Debug = "ignore")]
-        fetcher: Arc<Fetcher<Types, S, P>>,
-    },
-    /// Callback when fetching the leaf in order to then look up something else.
-    Continuation {
-        callback: HeaderCallback<Types, S, P>,
-    },
-}
-
-impl<Types: NodeType, S, P> PartialEq for LeafCallback<Types, S, P> {
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other).is_eq()
-    }
-}
-
-impl<Types: NodeType, S, P> Eq for LeafCallback<Types, S, P> {}
-
-impl<Types: NodeType, S, P> Ord for LeafCallback<Types, S, P> {
-    fn cmp(&self, other: &Self) -> Ordering {
-        match (self, other) {
-            // Store leaves in the database before storing derived objecs.
-            (Self::Leaf { .. }, Self::Continuation { .. }) => Ordering::Less,
-            (Self::Continuation { .. }, Self::Leaf { .. }) => Ordering::Greater,
-
-            (Self::Continuation { callback: cb1 }, Self::Continuation { callback: cb2 }) => {
-                cb1.cmp(cb2)
-            }
-            _ => Ordering::Equal,
-        }
-    }
-}
-
-impl<Types: NodeType, S, P> PartialOrd for LeafCallback<Types, S, P> {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl<Types: NodeType, S, P> Callback<LeafQueryData<Types>> for LeafCallback<Types, S, P>
-where
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types> + 'static,
-    P: AvailabilityProvider<Types>,
-{
-    async fn run(self, leaf: LeafQueryData<Types>) {
-        match self {
-            Self::Leaf { fetcher } => {
-                let height = leaf.height();
-                tracing::info!("fetched leaf {height}");
-                let mut storage = fetcher.storage.write().await;
-                if let Err(err) = store_leaf(&mut *storage, leaf).await {
-                    // It is unfortunate if this fails, but we can still proceed by
-                    // returning the leaf that we fetched, keeping it in memory.
-                    // Simply log the error and move on.
-                    tracing::warn!("failed to store fetched leaf {height}: {err}");
-                }
-            }
-            Self::Continuation { callback } => callback.run(leaf.leaf.block_header),
-        }
-    }
-}
-
-#[derive(Derivative)]
-#[derivative(Debug(bound = ""))]
-enum HeaderCallback<Types, S, P>
-where
-    Types: NodeType,
-{
-    /// Callback when fetching the leaf in order to then look up the corresponding block.
-    Payload {
-        #[derivative(Debug = "ignore")]
-        fetcher: Arc<Fetcher<Types, S, P>>,
-    },
-    /// Callback when fetching the leaf in order to then look up the corresopnding VID common data.
-    VidCommon {
-        #[derivative(Debug = "ignore")]
-        fetcher: Arc<Fetcher<Types, S, P>>,
-    },
-}
-
-impl<Types: NodeType, S, P> PartialEq for HeaderCallback<Types, S, P> {
-    fn eq(&self, other: &Self) -> bool {
-        self.cmp(other).is_eq()
-    }
-}
-
-impl<Types: NodeType, S, P> Eq for HeaderCallback<Types, S, P> {}
-
-impl<Types: NodeType, S, P> PartialOrd for HeaderCallback<Types, S, P> {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl<Types: NodeType, S, P> Ord for HeaderCallback<Types, S, P> {
-    fn cmp(&self, other: &Self) -> Ordering {
-        // Arbitrarily, we choose to run payload callbacks before VID callbacks.
-        match (self, other) {
-            (Self::Payload { .. }, Self::VidCommon { .. }) => Ordering::Less,
-            (Self::VidCommon { .. }, Self::Payload { .. }) => Ordering::Greater,
-            _ => Ordering::Equal,
-        }
-    }
-}
-
-impl<Types, S, P> HeaderCallback<Types, S, P>
-where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types> + 'static,
-    P: AvailabilityProvider<Types>,
-{
-    fn fetcher(&self) -> Arc<Fetcher<Types, S, P>> {
-        match self {
-            Self::Payload { fetcher } => fetcher.clone(),
-            Self::VidCommon { fetcher } => fetcher.clone(),
-        }
-    }
-
-    fn run(self, header: Header<Types>) {
-        match self {
-            Self::Payload { fetcher } => {
-                tracing::info!(
-                    "fetched leaf {}, will now fetch payload",
-                    header.block_number()
-                );
-                fetch_block_with_header(fetcher, header);
-            }
-            Self::VidCommon { fetcher } => {
-                tracing::info!(
-                    "fetched leaf {}, will now fetch VID common",
-                    header.block_number()
-                );
-                fetch_vid_common_with_header(fetcher, header);
-            }
-        }
-    }
-}
-
-async fn fetch_header_and_then<Types, S, P>(
-    storage: &RwLockReadGuard<'_, NotifyStorage<Types, S>>,
-    req: BlockRequest<Types>,
-    callback: HeaderCallback<Types, S, P>,
-) where
-    Types: NodeType,
-    Payload<Types>: QueryablePayload,
-    S: AvailabilityStorage<Types> + 'static,
-    P: AvailabilityProvider<Types>,
-{
-    match req {
-        BlockRequest::Id(id) => {
-            // Check if at least the header is available in local storage. If it is, we benefit
-            // two ways:
-            // 1. We know for sure the corresponding block exists, so we can unconditionally
-            //    trigger an active fetch without unnecessarily bothering our peers.
-            // 2. We only need to fetch the missing data (e.g. payload or VID common), not the full
-            //    block. Not only is this marginally less data to download, there are some providers
-            //    that may only be able to provide certain data. For example, the HotShot DA
-            //    committee members may be able to provide paylaods, but not full blocks. Or, in the
-            //    case where VID recovery is needed, the VID common data may be available but the
-            //    full block may not exist anywhere.
-            if let Some(header) = storage
-                .storage
-                .get_header(id)
-                .await
-                .context(format!("loading header for block {id}"))
-                .ok_or_trace()
-            {
-                callback.run(header);
-                return;
-            }
-
-            // If the header is _not_ present, we may still be able to fetch the request, but we
-            // need to fetch the header (in fact, the entire leaf) first. This is because we have an
-            // invariant that we should not store derived objects in the database unless we already
-            // have the corresponding header and leaf.
-            match id {
-                BlockId::Number(n) => {
-                    fetch_leaf_with_callbacks(callback.fetcher(), n.into(), [callback.into()]);
-                }
-                BlockId::Hash(h) => {
-                    // Given only the hash, we cannot tell if the corresonding leaf actually exists,
-                    // since we don't have a corresponding header. Therefore, we will not spawn an
-                    // active fetch.
-                    tracing::debug!("not fetching unknown block {h}");
-                }
-                BlockId::PayloadHash(h) => {
-                    // Same as above, we don't fetch a block with a payload that is not known to
-                    // exist.
-                    tracing::debug!("not fetching block with unknown payload {h}");
-                }
-            }
-        }
-        BlockRequest::WithTransaction(h) => {
-            // We don't actively fetch blocks when requested by transaction, because without the
-            // block payload, we have no way of knowing whether a block with such a transaction
-            // actually exists, and we don't want to bother peers with requests for non-existant
-            // blocks.
-            tracing::debug!("not fetching block with unknown transaction {h}");
-        }
-    }
 }
 
 /// Break a range into fixed-size chunks.
